bug-1884041: support java_exception in crash_report_to_description
This adds support for java_exception field when generating the
description for creating new bugs.

This also creates `webapp/crashstats/tests/test_libbugzilla.py` and
moves a bunch of tests from `test_jinja_helpers.py` so that
`test_libbugzilla.py` tests all the requirements for the functions.

I also switched the comment-building tests to use dedent instead of
string delimiters, newlines, and such. It's a little easier to
manipulate and a lot easier to read and verify.

This also re-does `test_corrected_os_version_name` so it's parametrized.